refactor(tools): decompose install_specific.py into per-subcommand package (#197)
Phase 9 of the Gen-3 anti-spiral plan: split the 721-line monolith into
a thin entry (47 lines) + a 16-file package under
tools/vm_scripts/install_specific/ where every file is ≤80 lines.

The package is already rsynced to /tmp/vm_scripts/ by the existing
stage_4/config_bundle.py; the thin entry resolves sys.path to either the
controller layout or the /tmp VM layout, then imports the cmd_* entries.
Pipeline stage scripts (stage_8 pre/post_restart_config.sh) invoke
`python3 /tmp/install_specific.py <subcommand>` unchanged.

Every function body is preserved verbatim — this is pure decomposition,
no behaviour change. verify_phase9.py enforces the size caps, stdlib-only
import constraint, and --help surface.

fixes #197
